requirements.txt
openai-whisper
imageio-ffmpeg
flask
yt-dlp
